# public/docs/openapi.yaml
openapi: 3.0.3
info:
  title: 'tax file backend docs'
  description: documentation
  version: 1.0.0
servers:
  -
    url: 'https://calm-spontaneity-production.up.railway.app'
paths:
...
  /api/auth/password-reset:
    post:
      summary: 'Send password reset email'
      operationId: sendPasswordResetEmail
      description: 'Send an email with password reset otp'
      parameters: []
      responses:
        200:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  message: 'password reset code sent successfully'
                  data: null
                properties:
                  message:
                    type: string
                    example: 'password reset code sent successfully'
                  data:
                    type: string
                    example: null
      tags:
        - 'Password Reset'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  description: 'Must be a valid email address.'
                  example: dubuque.clemens@example.com
                  nullable: false
              required:
                - email
      security: []
...
components:
  securitySchemes:
    default:
      type: http
      scheme: bearer
      description: 'You can retrieve your token by visiting your dashboard and clicking <b>Generate API token</b>.'
security:
  -
    default: []